Decomposed src/index.ts to src/server/*
